renamed directories from `host_` to `endpoint_`
