+++ x-pack/plugins/security_solution/public/common/lib/endpoint_isolation/index.ts
