+++ x-pack/plugins/security_solution/public/common/lib/endpoint_pending_actions/index.ts
@@ -5,4 +5,4 @@
  * 2.0.
  */
 
-export * from './host_pending_actions';
+export * from './endpoint_pending_actions';
